+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWRS1F89SXRHE1TZWAH21GH3"
+        "01KWTVR58FDB6F1B08GREEJJ0E"
     )
     r"""The unique identifier of the data source"""