+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWRS1F8ERQ5KCWA01G2VWP16"
+        "01KWTVR58MQ3WN60HTPSF8YNWM"
     )
     r"""The unique identifier of the data source"""