+++ packages/orq-rc/RELEASES.md
@@ -9299,3 +9299,13 @@ Based on:
 - [python v4.12.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.1 - packages/orq-rc
+
+## 2026-07-06 04:44:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.788.1 (2.915.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.2 - packages/orq-rc